+++ src/__tests__/extensions/replay/sessionrecording.test.ts
@@ -31,7 +31,17 @@ import {
 } from '../../../extensions/replay/sessionrecording'
 import { assignableWindow, window } from '../../../utils/globals'
 import { RequestRouter } from '../../../utils/request-router'
-import { customEvent, EventType, eventWithTime, pluginEvent } from '@rrweb/types'
+import {
+    customEvent,
+    EventType,
+    eventWithTime,
+    fullSnapshotEvent,
+    incrementalSnapshotEvent,
+    IncrementalSource,
+    metaEvent,
+    pluginEvent,
+} from '@rrweb/types'
+import { loadScript } from '../../../utils'
 import Mock = jest.Mock
 
 // Type and source defined here designate a non-user-generated recording event
@@ -42,8 +52,6 @@ jest.mock('../../../utils', () => ({
 }))
 jest.mock('../../../config', () => ({ LIB_VERSION: 'v0.0.1' }))
 
-import { loadScript } from '../../../utils'
-
 const loadScriptMock = loadScript as jest.Mock
 
 const EMPTY_BUFFER = {
@@ -53,19 +61,32 @@ const EMPTY_BUFFER = {
     windowId: null,
 }
 
-const createMetaSnapshot = (event = {}) => ({
-    type: META_EVENT_TYPE,
-    data: {},
-    ...event,
-})
+const createMetaSnapshot = (event = {}): metaEvent =>
+    ({
+        type: META_EVENT_TYPE,
+        data: {
+            href: 'https://has-to-be-present-or-invalid.com',
+        },
+        ...event,
+    } as metaEvent)
 
-const createFullSnapshot = (event = {}) => ({
-    type: FULL_SNAPSHOT_EVENT_TYPE,
-    data: {},
-    ...event,
-})
+const createStyleSnapshot = (event = {}): incrementalSnapshotEvent =>
+    ({
+        type: INCREMENTAL_SNAPSHOT_EVENT_TYPE,
+        data: {
+            source: IncrementalSource.StyleDeclaration,
+        },
+        ...event,
+    } as incrementalSnapshotEvent)
 
-const createIncrementalSnapshot = (event = {}) => ({
+const createFullSnapshot = (event = {}): fullSnapshotEvent =>
+    ({
+        type: FULL_SNAPSHOT_EVENT_TYPE,
+        data: {},
+        ...event,
+    } as fullSnapshotEvent)
+
+const createIncrementalSnapshot = (event = {}): incrementalSnapshotEvent => ({
     type: INCREMENTAL_SNAPSHOT_EVENT_TYPE,
     data: {
         source: 1,
@@ -895,6 +916,47 @@ describe('SessionRecording', () => {
             expect(sessionRecording.started).toEqual(false)
         })
 
+        it('can emit when there are circular references', () => {
+            sessionRecording.afterDecideResponse(makeDecideResponse({ sessionRecording: { endpoint: '/s/' } }))
+            sessionRecording.startIfEnabledOrStop()
+
+            const someObject = { emit: 1 }
+            // the same object can be there multiple times
+            const circularObject: Record<string, any> = { emit: someObject, again: someObject }
+            // but a circular reference will be replaced
+            circularObject.circularReference = circularObject
+            _emit(createFullSnapshot(circularObject))
+
+            expect(sessionRecording['buffer']).toEqual({
+                data: [
+                    {
+                        again: {
+                            emit: 1,
+                        },
+                        circularReference: {
+                            again: {
+                                emit: 1,
+                            },
+                            // the circular reference is captured to the buffer,
+                            // but it didn't explode when estimating size
+                            circularReference: expect.any(Object),
+                            emit: {
+                                emit: 1,
+                            },
+                        },
+                        data: {},
+                        emit: {
+                            emit: 1,
+                        },
+                        type: 2,
+                    },
+                ],
+                sessionId: sessionId,
+                size: 149,
+                windowId: 'windowId',
+            })
+        })
+
         describe('console logs', () => {
             it('if not enabled, plugin is not used', () => {
                 posthog.config.enable_recording_console_log = false
@@ -1236,7 +1298,7 @@ describe('SessionRecording', () => {
             _addCustomEvent.mockClear()
         })
 
-        it('does not emit when idle', () => {
+        it('does not emit plugin events when idle', () => {
             const emptyBuffer = {
                 ...EMPTY_BUFFER,
                 sessionId: sessionId,
@@ -1247,11 +1309,41 @@ describe('SessionRecording', () => {
             sessionRecording['isIdle'] = true
             // buffer is empty
             expect(sessionRecording['buffer']).toEqual(emptyBuffer)
-            // a plugin event doesn't count as returning from idle
-            sessionRecording.onRRwebEmit(createPluginSnapshot({}) as unknown as eventWithTime)
 
-            // buffer is still empty
+            sessionRecording.onRRwebEmit(createPluginSnapshot({}) as eventWithTime)
+
+            // a plugin event doesn't count as returning from idle
+            expect(sessionRecording['isIdle']).toEqual(true)
+            expect(sessionRecording['buffer']).toEqual({
+                ...EMPTY_BUFFER,
+                sessionId: sessionId,
+                windowId: 'windowId',
+            })
+        })
+
+        it('active incremental events return from idle', () => {
+            const emptyBuffer = {
+                ...EMPTY_BUFFER,
+                sessionId: sessionId,
+                windowId: 'windowId',
+            }
+
+            // force idle state
+            sessionRecording['isIdle'] = true
+            // buffer is empty
             expect(sessionRecording['buffer']).toEqual(emptyBuffer)
+
+            sessionRecording.onRRwebEmit(createIncrementalSnapshot({}) as eventWithTime)
+
+            // an incremental event counts as returning from idle
+            expect(sessionRecording['isIdle']).toEqual(false)
+            // buffer contains event allowed when idle
+            expect(sessionRecording['buffer']).toEqual({
+                data: [createFullSnapshot(), createIncrementalSnapshot({})],
+                sessionId: sessionId,
+                size: 50,
+                windowId: 'windowId',
+            })
         })
 
         it('emits custom events even when idle', () => {
@@ -1264,7 +1356,7 @@ describe('SessionRecording', () => {
                 windowId: 'windowId',
             })
 
-            sessionRecording.onRRwebEmit(createCustomSnapshot({}) as unknown as eventWithTime)
+            sessionRecording.onRRwebEmit(createCustomSnapshot({}) as eventWithTime)
 
             // custom event is buffered
             expect(sessionRecording['buffer']).toEqual({
@@ -1279,6 +1371,88 @@ describe('SessionRecording', () => {
                 ],
                 sessionId: sessionId,
                 size: 47,
+                windowId: 'windowId',
+            })
+        })
+
+        it('emits full snapshot events even when idle', () => {
+            // force idle state
+            sessionRecording['isIdle'] = true
+            // buffer is empty
+            expect(sessionRecording['buffer']).toEqual({
+                ...EMPTY_BUFFER,
+                sessionId: sessionId,
+                windowId: 'windowId',
+            })
+
+            sessionRecording.onRRwebEmit(createFullSnapshot({}) as eventWithTime)
+
+            // custom event is buffered
+            expect(sessionRecording['buffer']).toEqual({
+                data: [
+                    {
+                        data: {},
+                        type: 2,
+                    },
+                ],
+                sessionId: sessionId,
+                size: 20,
+                windowId: 'windowId',
+            })
+        })
+
+        it('emits meta snapshot events even when idle', () => {
+            // force idle state
+            sessionRecording['isIdle'] = true
+            // buffer is empty
+            expect(sessionRecording['buffer']).toEqual({
+                ...EMPTY_BUFFER,
+                sessionId: sessionId,
+                windowId: 'windowId',
+            })
+
+            sessionRecording.onRRwebEmit(createMetaSnapshot({}) as eventWithTime)
+
+            // custom event is buffered
+            expect(sessionRecording['buffer']).toEqual({
+                data: [
+                    {
+                        data: {
+                            href: 'https://has-to-be-present-or-invalid.com',
+                        },
+                        type: 4,
+                    },
+                ],
+                sessionId: sessionId,
+                size: 69,
+                windowId: 'windowId',
+            })
+        })
+
+        it('emits style snapshot events even when idle', () => {
+            // force idle state
+            sessionRecording['isIdle'] = true
+            // buffer is empty
+            expect(sessionRecording['buffer']).toEqual({
+                ...EMPTY_BUFFER,
+                sessionId: sessionId,
+                windowId: 'windowId',
+            })
+
+            sessionRecording.onRRwebEmit(createStyleSnapshot({}) as eventWithTime)
+
+            // custom event is buffered
+            expect(sessionRecording['buffer']).toEqual({
+                data: [
+                    {
+                        data: {
+                            source: 13,
+                        },
+                        type: 3,
+                    },
+                ],
+                sessionId: sessionId,
+                size: 31,
                 windowId: 'windowId',
             })
         })
@@ -1358,7 +1532,6 @@ describe('SessionRecording', () => {
             )
 
             // this triggers exit from idle state _and_ is a user interaction, so we take a full snapshot
-
             const fourthSnapshot = emitActiveEvent(fourthActivityTimestamp)
             expect(_addCustomEvent).toHaveBeenCalledWith('sessionNoLongerIdle', {
                 reason: 'user activity',